fix: resolve three mypy type errors (CI green)
- brain/redactor.py: rename NamedTuple field count → occurrences;
  tuple.count() is a built-in method so mypy flagged the int assignment
  as incompatible with Callable[[Any], int] (updated all 3 usage sites)
- analytics/siem.py: type hec_event as dict[str, Any] so float timestamp
  can be assigned to the "time" key without mypy incompatible-type error
- analytics/logger.py: widen days signature int|None → float|None so the
  dashboard can pass 1/24 (1 hour) without a type mismatch

+++ warden/analytics/logger.py
@@ -70,7 +70,7 @@ def append(entry: dict) -> None:
 
 # ── Reader (used by the dashboard) ───────────────────────────────────────────
 
-def load_entries(days: int | None = None) -> list[dict]:
+def load_entries(days: float | None = None) -> list[dict]:
     """
     Read all log entries, optionally filtering to the last *days* days.
     Returns an empty list if the file does not exist yet.
